split files; added validation generation file
