+++ src/Python/validation/verify_calculations.py
@@ -1,94 +1,9 @@
-from cgi import test
-import numpy as np
 import pandas as pd
 
-import wows_shell
+import generate_data
+from validation_files_util import read_reference_csvs
 
-reference_directory = 'verification_reference'
-
-def generate_csv(df, filename: str):
-    with open(filename, 'w') as f:
-        df.to_csv(f, index=False, line_terminator='\n')
-
-def generate_reference_csvs(impact_dataframe, angles_dataframe, dispersion_dataframe, post_penetration_dataframe):
-    generate_csv(impact_dataframe, f'{reference_directory}/impact_data.csv')
-    generate_csv(angles_dataframe, f'{reference_directory}/angles_data.csv')
-    generate_csv(dispersion_dataframe, f'{reference_directory}/dispersion_data.csv')
-    generate_csv(post_penetration_dataframe, f'{reference_directory}/postpenetration_data.csv')
-
-def generate_ordered_enum_list(enum_dict):
-    ordered = [None] * len(enum_dict)
-    for name, data in enum_dict.items():
-        ordered[int(data)] = name
-    
-    return ordered
-
-def read_reference_csvs():
-    return (
-        pd.read_csv(f'{reference_directory}/impact_data.csv'),
-        pd.read_csv(f'{reference_directory}/angles_data.csv'),
-        pd.read_csv(f'{reference_directory}/dispersion_data.csv'),
-        pd.read_csv(f'{reference_directory}/postpenetration_data.csv')
-    )
-
-test_shell = wows_shell.shell(
-    wows_shell.shellParams(.460, 780, .292, 1460, 2574, 6, .033, 76, 45, 60, 0),
-    wows_shell.dispersionParams(10, 2.8, 1000, 5000, 0.5, 0.2, 0.6, 0.8, 26630, 2.1),
-    "Yamato")
-
-calculator = wows_shell.shellCalc()
-calculator.setMax(25)
-calculator.setMin(0)
-calculator.setPrecision(0.1)
-calculator.setX0(0)
-calculator.setY0(0)
-calculator.setDtMin(0.02)
-
-calculator.calcImpactForwardEuler(test_shell, addTraj=True)
-calculator.calcAngles(test_shell, 410, -15)
-calculator.calcDispersion(test_shell, int(wows_shell.verticalTypes.horizontal))
-angles = [0, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60]
-calculator.calcPostPen(test_shell, 70, 0, angles, True, False)
-
-impact_data = test_shell.getImpact()
-#print(impact_data.shape)
-
-impact_dataframe = pd.DataFrame(impact_data.transpose(), 
-    columns=generate_ordered_enum_list(wows_shell.impactIndices.__members__))
-
-#print(impact_dataframe)
-
-angles_data = test_shell.getAngles()
-#print(angles_data.shape)
-
-angles_dataframe = pd.DataFrame(angles_data.transpose(),
-    columns=generate_ordered_enum_list(wows_shell.angleIndices.__members__))
-
-#print(angles_dataframe)
-
-dispersion_data = test_shell.getDispersion()
-#print(dispersion_data.shape)
-
-dispersion_dataframe = pd.DataFrame(dispersion_data.transpose(),
-    columns=generate_ordered_enum_list(wows_shell.dispersionIndices.__members__))
-
-#print(dispersion_dataframe)
-
-post_penetration_data = test_shell.getPostPen()
-post_penetration_data_shape = post_penetration_data.shape
-#print(post_penetration_data_shape)
-#print(post_penetration_data)
-
-post_penetration_dataframe = pd.DataFrame(
-    post_penetration_data
-        .reshape((post_penetration_data_shape[0], post_penetration_data_shape[1] * post_penetration_data_shape[2]))
-        .transpose(),
-    columns=generate_ordered_enum_list(wows_shell.postPenIndices.__members__))
-#print(post_penetration_dataframe)
-
-#generate_reference_csvs(impact_dataframe, angles_dataframe, dispersion_dataframe, post_penetration_dataframe)
-
-#Validation Phase
+impact_dataframe, angles_dataframe, dispersion_dataframe, post_penetration_dataframe = generate_data.get_dataframes()
 
 impact_dataframe_reference, angles_dataframe_reference, dispersion_dataframe_reference, post_penetration_dataframe_reference = read_reference_csvs()
